Root CA and library added
